Add WTF and bmalloc modulemaps https://bugs.webkit.org/show_bug.cgi?id=292791 rdar://151023888
Reviewed by Elliott Williams.

This commit adds modulemap files for WTF and bmalloc. These are necessary
because Swift/C++ interop interprets C++ headers using the "-fmodules" option,
interpreting the C++ codebase as clang modules.

Ultimately this PR results in just two new module.modulemap
files alongside the headers. However, this can cause build time
differences in all the transitive dependencies of WTF and bmalloc
because they'll interpret header files differently.

For now, this PR just puts such modulemaps in place for OS X 26+.
Future PRs will extend the modulemaps into other targets.

The approach taken for the two modulemaps is different.

* bmalloc has a hard-coded modulemap, defining a single monolithic module.
* WTF has an autogenerated modulemap, where each header typically
  resides in a small submodule.

The approach of many small submodules was previously used in
WebKit_Internal, and is used here because Swift/C++ interoperability otherwise
tries to interpret the whole module, and that results in errors such as:

  /usr/include/c++/v1/__memory/construct_at.h:47:58: note: candidate template ignored: substitution failure [with _Tp = WTF::TextBreakIterator, _Args = <const WTF::TextBreakIterator &>]: call to deleted constructor of 'WTF::TextBreakIterator'
_LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX20 _Tp* __construct_at(_Tp* __location, _Args&&... __args) {

This is presumed to be rdar://148760413 or related overinstantiation
of templates. We'll likely need to solve those in future, but meanwhile
splitting the module into lots of submodules limits the blast radius of
such issues and allows us to make progress with Swift.

Why is one manually generated and the other autogenerated?
Per the comments in the bmalloc modulemap, we only want to expose
a hand-curated set of headers in the bmalloc module, and therefore
there's no benefit to autogenerating it. It's specifically the
dependency on a certain value of OS_UNFAIR_LOCK_INLINE which
forces this decision, and that's discussed in
rdar://153287512.

Whereas, for WTF we want to expose all the headers, and thus evolving the
modulemap to add new headers and remove older headers would be error-prone and
time consuming - hence we autogenerate.

For the WTF mdoulemap, we use a script and a TOML file that describes the
quirks of given header files (for example, whether they're interdependent or
can be used in plain C contexts). See the TOML files in this PR and their
comments for a description of these quirks. Thanks to Elliott Williams for
writing the script.

The script converts a tapi filelist-style header listing
into a modulemap, supporting both monolithic and submodule layouts.
This process works well for libraries like WTF which have no
umbrella header, and which have no concept of public vs private API.

We do not export the modulemaps to the build output on pre-OS X 15;
doing so confuses the build of PALSwift.

This commit also adjusts some headers to be more compatible with
modules, adding missing includes, and ensuring all includes are
outside of any extern "C" blocks.

This PR is the third attempt at landing modulemaps for WTF and bmalloc.
The first time didn't work because it turned out that we needed to resolve
three issues first, to unbreak downstream consumers:

* We needed to move away from reliance on UCHAR_TYPE in WTF.
  (rdar://153817245).
* Downstream consumers of WebKit were previously using some WTF definitions
  without including WTF headers. That ceases to work when WTF becomes a
  module; we have removed that dependency (rdar://154081231)
* #45625 resolves an issue with AppServerSupport which starts to appear
  once a modulemap is in place. There were also small fixes to the
  modularization of various system dependencies included within WebKit.
  (Because some of these fixes in WebKit's dependencies have only landed
  in recent SDKs, we do not export the modulemaps when older SDKs
  are in use.)

The second time didn't work because we needed to:
* Be robust to different versions of the python tomli module, some of
  which require binary input and some of which require text input.
* Not attempt to use modules in downlevel builds. (Previously we were
  checking the XCode version; we should have checked the target SDK
  version and that's what we now do).

* Source/WTF/Configurations/WTF.xcconfig:
* Source/WTF/Configurations/modulemap.toml: Added.
* Source/WTF/Scripts/modulemap-from-tapi-filelist.py: Added.
* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/Compiler.h:
* Source/WebCore/PAL/pal/spi/cg/CoreGraphicsSPI.h:
* Source/WebKit/WebKitSwift/IdentityDocumentServices/WKIdentityDocumentPresentmentDelegate.h:
* Source/bmalloc/Configurations/bmalloc.xcconfig:
* Source/bmalloc/Configurations/module.modulemap: Added.
* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj:

Co-authored-by: Elliott Williams <emw@apple.com>
Canonical link: https://commits.webkit.org/297469@main