Deprecate FastMalloc macro for new types https://bugs.webkit.org/show_bug.cgi?id=295749 rdar://155564529
Reviewed by Devin Rousso.

As part of our ongoing effort to move over to TZones, we want to make it
obvious to developers that the WTF_MAKE_FAST_ALLOCATED macro-family is
not to be used except when necessary. By adding "DEPRECATED" to the
name, this patch should hopefully give people second thoughts on using
the macro.

Developers should prefer to use the WTF_MAKE_TZONE_ALLOCATED family of
macros whenever possible; broadly, we expect them to be equi-performant
while providing security benefits beyond FastMalloc.
However, since TZone does not handle variable-sized types, this may not
be possible everywhere. If it is not possible to use TZone for a given
type, or if migrating it over shows a measurable performance regression,
then FastMalloc can be used for the time being -- it is still preferable
to using the system allocator.

Beyond that, this also changes the _FAST_ macro family to take the
type-name as a parameter: this should make it easier to migrate the
remaining objects when we do get around to it.

Canonical link: https://commits.webkit.org/297491@main