[Style] Remove derived value accessors for grid-template-rows/grid-template-columns from RenderStyle https://bugs.webkit.org/show_bug.cgi?id=295987
Reviewed by Alan Baradlay.

There are a large number of derived value accessors for `grid-template-rows`/`grid-template-columns`
in RenderStyle that simply forward to the underlying member. In an effort to reduce the size of the
RenderStyle interface, these can be easily replaced by inlining their contents.

In a followup, these call sites will be further reduced by utilizing the `gridTemplateList` function
that takes a direction.

* Source/WebCore/inspector/InspectorOverlay.cpp:
* Source/WebCore/rendering/GridTrackSizingAlgorithm.cpp:
* Source/WebCore/rendering/RenderBlock.cpp:
* Source/WebCore/rendering/RenderBox.cpp:
* Source/WebCore/rendering/RenderGrid.cpp:
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/RenderStyleInlines.h:
* Source/WebCore/rendering/style/StyleGridData.h:
* Source/WebCore/style/StyleExtractorCustom.h:
* Source/WebCore/style/StyleOrderedNamedLinesCollector.h:

Canonical link: https://commits.webkit.org/297503@main