Extend timeout for safe browsing check https://bugs.webkit.org/show_bug.cgi?id=296079 rdar://155478524
Reviewed by Wenson Hsieh.

In 294690@main we added a timeout so that if a safe browsing provider does not respond
in a timely manner we proceed with the load anyways.  This will hopefully not happen in
practice because a safe browsing provider typically has low latency, but we don't want
a safe browsing provider's failure to block loading completely.

The timeout we added was the amount of time the request took to roundtrip to a response
beyond the time the response was received.  Most of the time a timeout of no additional
time is needed, but on some fast sites this is occasionally not enough time.  To be a
little more generous on the amount of additional time we are willing to wait, wait up to
1.5x the roundtrip time plus a small constant in case the roundtrip time is so close to
zero that even multiplying it doesn't do much.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::decidePolicyForResponseShared):

Canonical link: https://commits.webkit.org/297504@main