com.apple.WebKit.GPU infinite recursion in  WebCore::VideoMediaSampleRenderer::decodeNextSampleIfNeeded https://bugs.webkit.org/show_bug.cgi?id=296035 rdar://154251293
Reviewed by Jer Noble.

Replace re-entrant call with a loop.
It is unclear how the original issue could happen in the first place as we would
always exit as soon as the compressed queue is empty, and there can never be
more than 50 compressed samples in the queue.

No change in behaviour, covered by existing tests.

Canonical link: https://commits.webkit.org/297517@main