Reviewed-by validator fails intermittently with long commit messages rdar://155614510 https://bugs.webkit.org/show_bug.cgi?id=296012
Reviewed by Elliott Williams.

Instead of piping the output of `git log`, save it to a file. This forces
the entire output of `git log` to be emitted and grep should return as expected, no
matter the length of the commit message.

This approach uses an actual file since process substitution is not supported
in POSIX.

* Tools/CISupport/ews-build/steps.py:
(ValidateCommitMessage.run):
* Tools/CISupport/ews-build/steps_unittest.py:

Canonical link: https://commits.webkit.org/297529@main