Add a quirk to disable DOMAudioSession on descript.com https://bugs.webkit.org/show_bug.cgi?id=296158 rdar://156024693
Reviewed by Jer Noble.

Descript uses the DOMAudioSession API to set the session type to "playback", but this has the side
effect of making the page's audio session not compatible with audio capture (concretely, the
promise returned by getUserMedia() is rejected with an exception and the user is never prompted for
microphone access). This is likely not the behavior the site intended when they adopted
DOMAudioSession, and since this is still an experimental API, this patch disables it on descript to
restore microphone capture functionality.

* Source/WebCore/Modules/audiosession/DOMAudioSession.idl:
* Source/WebCore/Modules/audiosession/Navigator+AudioSession.idl:
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldDisableDOMAudioSessionQuirk const):
(WebCore::handleDescriptQuirks):
(WebCore::Quirks::determineRelevantQuirks):
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/QuirksData.h:

Canonical link: https://commits.webkit.org/297571@main