Use std::unique_ptr instead of MallocPtr in ConcurrentBuffer and SegmentedVector https://bugs.webkit.org/show_bug.cgi?id=296166 rdar://156121020
Reviewed by Keith Miller.

We are about to harden MallocPtr by requiring the parameter type to be trivially
destructible. That will prevent accidental memory leaks, when a type is used with
MallocPtr under the assumption it does not need a destructor call, but is later changed to
require non-trivial destruction (e.g. rdar://152574864).

So that MallocPtr can be hardened as intended, this patch changes two recently added
usages of MallocPtr to instead use std::unique_ptr with a deleter that disposes of the
pointer the same way MallocPtr would.

Canonical link: https://commits.webkit.org/297601@main