Add missing includes https://bugs.webkit.org/show_bug.cgi?id=296125 rdar://156057537
Reviewed by Richard Robinson

This adds some includes to make each of these header files self-contained,
which is necessary when we build them as clang modules. This will happen once
we enable Swift/C++ interop.

This is a reland without a change to TZoneMalloc.h, which triggered build
problems in modules within Safari. We can land that separately later as
necessary.

* Source/JavaScriptCore/runtime/BufferMemoryHandle.h:
* Source/WTF/wtf/CheckedPtr.h:
* Source/WTF/wtf/RefPtr.h:
* Source/WTF/wtf/WeakPtrFactory.h:
* Source/WebCore/platform/RectEdges.h:
* Source/WebCore/platform/graphics/ColorMatrix.h:
* Source/WebCore/platform/graphics/IntSize.h:

Canonical link: https://commits.webkit.org/297676@main