REGRESSION(291932@main): App may crash when a page calls ApplePaySetup.begin() https://bugs.webkit.org/show_bug.cgi?id=296122 rdar://155350256
Reviewed by Alex Christensen.

In 291932@main, we introduced a CoreIPC wrapper for
PKPaymentSetupFeature and introduced a web process assertion since we do
not want NSKU usage outside of the web process. While this still holds
true, it turns out that ApplePaySetup.begin() produces an IPC message to
WebPaymentCoordinatorProxy with a message containing
PKPaymentSetupFeature. This means the Apple Pay API in question easily
crashes the proxy hosting process (either UI or Networking).

In this patch, we make the codepath less crash happy by removing the
process assertion, calling into NSKU only while we're in the web
process, and simply returning `nil` if we are in some other process.

* Source/WebKit/Shared/Cocoa/CoreIPCPKPaymentSetupFeature.mm:
(WebKit::CoreIPCPKPaymentSetupFeature::toID const):

Canonical link: https://commits.webkit.org/297686@main