[visionOS] In fullscreen, unable to change subtitles after changing playback speed https://bugs.webkit.org/show_bug.cgi?id=296223 rdar://154874849
Reviewed by Jer Noble.

When the user changes a legible or audio track in LMPlayableViewController, LinearMediaKit calls
setLegibleTrack(_:) or setAudioTrack(_:) on WebKit's Playable-conforming object. We expect the
Track passed to those functions to match (by object identity) one of the tracks we had previously
published in legibleTracksPublisher or audioTracksPublisher, but due to a bug that is not always
the case. When the Track object can't be found in the corresponding published array, we fail to
find an index for the new track and consequently fail to activate it.

Worked around this by falling back to matching the track by its localized display name when we
cannot match it by object identity.

* Source/WebKit/Platform/ios/PlaybackSessionInterfaceLMK.mm:
(-[WKLinearMediaPlayerDelegate linearMediaPlayer:setAudioTrack:]):
(-[WKLinearMediaPlayerDelegate linearMediaPlayer:setLegibleTrack:]):

Canonical link: https://commits.webkit.org/297688@main