Rename computeIntrinsicRatioInformation to computeIntrinsicSizeAndPreferredAspectRatio. https://bugs.webkit.org/show_bug.cgi?id=296284 rdar://156333988
Reviewed by Brandon Stewart and Brent Fulgham.

computeIntrinsicRatioInformation is a pretty vague name so let's rename
it to be a bit more descriptive. Since this function seems to have two
primary goals we can rename it computeIntrinsicSizeAndPreferredAspectRatio.
Although this is a bit more wordy, this does a better job of capturing
what this function does.

Rename the intrinsicRatio argument also to preferredAspectRatio since
that is what these functions are actually computing by checking the
aspect-ratio property and not just returning the ratio associated with
the replaced content.

Canonical link: https://commits.webkit.org/297705@main