Apply `_setWebViewTransform3DForModel` to WKContentView https://bugs.webkit.org/show_bug.cgi?id=296292 rdar://154223351
Reviewed by Wenson Hsieh.

This PR addresses an issue on visionOS where previously applying the transform3D on
the scrollView or an ancestor of a scrollView could affect the gaze effect on the
scroll indicator. To remedy this, we will apply the transform3D onto the content view
instead which is a descendant of the scroll view.

* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.h:
* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.mm:
(-[WKWebView scrollViewDidZoom:]):
(-[WKWebView _willInvalidateDraggedModelWithContainerView:]):
(-[WKWebView _setWebViewTransform3DForModel:]): Deleted.
* Source/WebKit/UIProcess/ios/WKContentView.h:
* Source/WebKit/UIProcess/ios/WKContentView.mm:
(-[WKContentView _setWebViewTransform3DForModel:]):

Canonical link: https://commits.webkit.org/297716@main