Enable modulemap on iPhone https://bugs.webkit.org/show_bug.cgi?id=296118 rdar://156040652
Reviewed by Richard Robinson.

This enables the production of module.modulemap files for the latest SDK on iOS
for WTF and bmalloc. These files describe how to interpret bmalloc and WTF as
clang modules, and that can influence the way downstream targets get built. We
are therefore enabling it platform-by-platform to give us time to react to
any unforeseen surprises.

* Configurations/WebKitTargetConditionals.xcconfig:
* Source/WTF/Configurations/WTF.xcconfig:
* Source/bmalloc/Configurations/bmalloc.xcconfig:

Canonical link: https://commits.webkit.org/297730@main