Use Copy File build phases instead of rsync to copy testapiScripts https://bugs.webkit.org/show_bug.cgi?id=296148 rdar://151903358
Reviewed by Yusuke Suzuki.

Internal bot failures reported in the radar were caused by the `testapiScripts` directory
not being included in some `build-root` configurations into the built artifacts. That was
because the scripts were copied using `rsync` into the location where `build-root` did not
expect them.

The patch changes the project to use the built-in `Copy Files` build phase.
`build-root` is aware of that phase and will include the copied files into the artifacts
as expected.

The downside of this solution is that new scripts added to that directory subtree need
to be manually added to the `Copy Files` build phases. I added a README to point that
out and explain the process.

Canonical link: https://commits.webkit.org/297740@main