[Writing Tools] Support semantic HTML lists as input to Writing Tools https://bugs.webkit.org/show_bug.cgi?id=296301 rdar://146968196
Reviewed by Wenson Hsieh.

When gathering the text context for Writing Tools, the resulting NSAttributedString needs
to have its relevant paragraph styles marked with NSTextLists so that Writing Tools can understand
that the content has lists in it and can preserve them.

To facilitate this, a new option is added to EditingHTMLConverter which enables extracting NSTextLists.
This is done by recursively traversing upwards from each element to find the containing list elements,
and then convert them to NSTextLists. The algorithm also employs caching and memoization to optimize
the amount of traversals needed.

* Source/WebCore/editing/cocoa/EditingHTMLConverter.h:

Add a `TextLists` option so that clients can opt-in to this new behavior.

* Source/WebCore/editing/cocoa/EditingHTMLConverter.mm:
(WebCore::enclosingElement):
(WebCore::enclosingLinkElement):
(WebCore::enclosingListElement):

Generalize the `enclosingLinkElement` function so that it can be reused for any arbitrary predicate,
in this case for determining if something is a list element or not.

(WebCore::associateElementWithTextLists):

The primary function used to do the extraction using a recursive algorithm; the output of the algorithm
is mutations to the `textListsForListElements` parameter. More details can be found in the inline comments.

(WebCore::updateAttributes):
(WebCore::editingAttributedStringInternal):

Modify the paragraph style of the attributes to account for the NSTextLists.

* Source/WebCore/page/writing-tools/WritingToolsController.mm:
(WebCore::WritingToolsController::willBeginWritingToolsSession):

Use the `TextLists` option.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WritingTools.mm:
(htmlToAttributedString):
(runTest):
(TEST(WritingToolsContextGeneration, ContextWithBasicOrderedList)):
(ContextWithCustomizedOrderedList)):
((WritingToolsContextGeneration, ContextWithNestedLists)):
((WritingToolsContextGeneration, ContextWithDiscontiguousLists)):
((WritingToolsContextGeneration, ContextWithBasicUnorderedList)):
((WritingToolsContextGeneration, ContextWithStyledContentChildrenInList)):

Add a lot of tests in a unique test suite for validation of the algorithm for different content.

* Tools/TestWebKitAPI/cocoa/DecomposedAttributedText.h: Added.
(_DecomposedAttributedTextBold::_DecomposedAttributedTextBold):
(_DecomposedAttributedTextItalic::_DecomposedAttributedTextItalic):
(_DecomposedAttributedTextOrderedList::_DecomposedAttributedTextOrderedList):
(_DecomposedAttributedTextUnorderedList::_DecomposedAttributedTextUnorderedList):
* Tools/TestWebKitAPI/cocoa/DecomposedAttributedText.mm: Added.
(convertMarker):
(decomposeChildren):
(decomposeAttributedTextRuns):
(runsForAttributedText):
(decompose):
(operator==):
(operator<<):

Add a testing mechanism to make it easier to compare HTML string converted to an NSAttributedString
value, and some expected value expressed as an HTML-like structure DSL.

Canonical link: https://commits.webkit.org/297769@main