[Navigation API] autofocus test fails because it falsely expects disabling to synchronously unfocus https://bugs.webkit.org/show_bug.cgi?id=296357 rdar://156452137
Reviewed by Tim Nguyen.

This test has a check that assumes that if an element becomes disabled, and it
was the active/focused element, that document.activeElement should immediately change.

The HTML spec does not enforce that if an element becomes disabled (and thus,
unfocusable), that the focused element must change immediately. So this check
is faulty. Chrome does appear to do this synchronoulsy, but Safari and Mozilla
do not, and so the test fails for them.

This check isn't actually checking the behavior of the Navigation API, but rather the
behavior of focus. Since the check is faulty, unrelated to the Navigation API, and is
causing the test to fail on 2/3 browsers, we remove it.

* LayoutTests/imported/w3c/web-platform-tests/navigation-api/focus-reset/autofocus-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/navigation-api/focus-reset/autofocus.html:

Canonical link: https://commits.webkit.org/297776@main