[visionOS] Immersive playback will sometimes get captions in an incorrect aspect ratio rdar://154211170 https://bugs.webkit.org/show_bug.cgi?id=296365
Reviewed by Eric Carlson.

When the TextTrackRepresentation is first created and given to the HTMLMediaElement, it has its
bounds initialized to the size (and therefore aspect ratio) of the underlying video box. For most
presentations, this is fine; desirable even! The captions are supposed to be overlayed atop the
video box. However, for immersive playback modes, the immersive presentation can display the
captions layer at a size and aspect ratio independent of the underlying video's naturalSize.

When we initialize the bounds of the TextTrackRepresentation, and we've already been given
the desired bounds of the TextTrackRepresentation, use those rather than the video display area.

* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::setTextTrackRepresentation):

Canonical link: https://commits.webkit.org/297796@main