[Absolute Positioning] Add simple end self alignment support for abspos child of a grid. https://bugs.webkit.org/show_bug.cgi?id=296345 rdar://156438264
Reviewed by Alan Baradlay.

In 297724@main we added some skeleton code for a new codepath which is
supposed to perform alignment of an absolutely positioned box within its
static position rectangle/range. This patch serves as a first step to
build up this path by adding some support when an abspos box is a child
of a grid. Although we relax the eligibility to allow grid content to
potentially run through this new path, we still maintain some strictness
by only allowing content in fairly simple and straightforward contexts.
These restrictions include:

1.) Making sure that the containing block is not the same as the parent.
This is because the definition of the static position rectangle is
different in this context for grids.

2.) Forcing the alignment subject, containing block, and parent to all
be horizontal-tb and LTR. Purely just to keep things simple.

3.) Only allowing end self alignment with default position type and
default overflow behavior which will get expanded on over time.

4.) Making sure that the alignment subject is smaller than the available
space since there is special logic for safety which will need to be
addressed in the future.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-end-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-end.html: Added.
Fairly straightforward tests to make sure this new functionality is
correct. Both contain grids with abspos children that are simply aligned
towards the end with lots of available space. One test has variable
border and padding to make sure everything is okay in that context.

* Source/WebCore/rendering/PositionedLayoutConstraints.cpp:
(WebCore::PositionedLayoutConstraints::isEligibleForStaticRangeAlignment const):
Modified to take in the remaining space in the static range along with
the size of the box. This is used to make a decision based on the
alignment safety and for now we only support content in which the box
fits within the available space.

(WebCore::PositionedLayoutConstraints::resolvePosition const):
We call into resolveAlignmentShift since it can compute the amount the
box needs to move to satisfy the alignment. Also, add a new lambda which
computes the size of the static range. Since this is a dependency on
eligibility, I opted to create it as a lambda here and not a static
function so that it cannot get called in contexts we do not support yet.
* Source/WebCore/rendering/PositionedLayoutConstraints.h:

Canonical link: https://commits.webkit.org/297803@main