[BBQ] Extract Register Allocator for Shared Use https://bugs.webkit.org/show_bug.cgi?id=296341 rdar://156433214
Reviewed by Dan Hecht.

This patch extracts BBQ's register allocator in to a shared header. There shouldn't be a significant
behavior change between the shared one and the new one. The new allocator takes a RegisterBank template
argument that defines the different details for both the JIT backend and register file in question.

The new class mostly follows the same terminology as BBQ did. The biggest change is that what was
called the LRU is now the Spiller, which operates on a spill hint. When we need to find a register
to evict we search for the value with the smallest hint. The highest bit is reserved for locking
a register or indicating that a register is ineligible. Additionally, the allocator has a abstract
RegisterBinding that it tracks and provides to the backend when requested or flushing. This binding
is intended to track any details the backend needs about the status of the register.

Lastly fix EnumTraits to work with Clang's safe buffer access for the enum's type name.

Canonical link: https://commits.webkit.org/297809@main