JavaScriptEvaluationResult's conversion to NSArrays should map null and undefined to NSNull https://bugs.webkit.org/show_bug.cgi?id=296401 rdar://156540084
Reviewed by Abrar Rahman Protyasha.

When returning a value from JS and converting it to ObjC, null maps to NSNull and undefined maps to nil.
However, when it is the content of a JS array, both map to NSNull.  This matches the behavior of a function
in JavaScriptCore named containerValueToObject that was implemented that way in 2013 in 124088@main.
Confusingly, conversion to NSDictionary does not have this behavior.
I accidentally change this in 297201@main which was fixed in rdar://156177377 by reverting it.
This brings the behavior to how it was for 12 years to maintain compatibility, and I add a test so
we won't miss it again in the future.

* Source/WebKit/Shared/JavaScriptEvaluationResult.mm:
(WebKit::JavaScriptEvaluationResult::toID):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebViewEvaluateJavaScript.mm:
(ReturnTypes)):

Canonical link: https://commits.webkit.org/297811@main