Fix and export flexbox multiline-align-self test https://bugs.webkit.org/show_bug.cgi?id=296414 rdar://156555222
Reviewed by Tim Nguyen.

It seems like this test was initially added in 97449@main as part of the
initial implementation of flex-wrap: wrap. It seems like Blink also has
had this test but updated the expectations over time for various
reasons. The test exercises the various values for align-self on flex
items in a multitude of contexts. Rather than going one by one, thinking
deeply about each subtest, and updating the expectations to what I think
are the right values, I am updating the expectations to match what Blink
currently has.

As part of this change, I also am exporting the test to WPT so that other
engines end up aware of any changes we make to the test and vice versa.
I also unrolled the generated content from the script and inserted the
resulting DOM tree in the test by just copying and pasting from the web
inspector. This will make it easier to view, debug, and make changes to
the test in the future. We could also break this test up into different
files for organizational purposes, but for now, just keep them in the same
file.

Canonical link: https://commits.webkit.org/297822@main