Run video-suppress-hdr-fullscreen.html again on iPad https://bugs.webkit.org/show_bug.cgi?id=296257 rdar://152860818
Reviewed by Cameron McCormack.

The test expectation was incorrect when run on iPad, for yet-unknown
reasons. This patch will attempt to reduce the chance of incorrect runs,
while adding some extra logging that should help if it happens again.

* LayoutTests/platform/ipad/TestExpectations:
Remove [failure] expectation so that the test is run again.

* LayoutTests/media/video-suppress-hdr-fullscreen.html:
- Don't force SupportHDRDisplayEnabled, so the test will only really
  work on recent platforms that do support HDR display functionality.
- Add some extra logging if the error condition happens again, to help
  with future debugging if that's needed.

Canonical link: https://commits.webkit.org/297844@main