Fix grid-abspos-staticpos-align-self-end-large-border-padding reference https://bugs.webkit.org/show_bug.cgi?id=296441 rdar://156634932
Reviewed by Vitor Roriz.

The item in this test reference should not be abspos since that is what
we are testing. It should just be a normal grid item so remove
position: absolute from it. grid-abspos-staticpos-align-self-end also
had a typo (missing px) in the border value so fix that.

Also rename the class from abspos to item so that is not misleading
anymore.

Canonical link: https://commits.webkit.org/297850@main