REGRESSION(294421@main): 2x http/wpt/mediarecorder (layout-tests) tests are constant text failures` https://bugs.webkit.org/show_bug.cgi?id=296467 rdar://156636980
Reviewed by Jer Noble, Eric Carlson, and Youenn Fablet.

In 294421@main we enabled asynchronouse video decoding via VTDecompressionSession.
It slightly changed the timing of operation and now an error logged on the console
always appeared in the test results causing it to fail.

The error occurred because when creating a VTCompressionSession we set the
key `kVTCompressionPropertyKey_AllowFrameReordering` even when encoding
h264 with a base profile. A base profile doesn't allow for b-frames as such
we don't need to set that key.

Covered by existing tests. No observable change.

* LayoutTests/platform/ios/TestExpectations: Re-enable test.
* Source/ThirdParty/libwebrtc/Source/webrtc/webkit_sdk/objc/components/video_codec/RTCVideoEncoderH264.mm:
(-[RTCVideoEncoderH264 configureCompressionSession]): Check if we are to use a base profile before setting the key.
We can't use `_useBaseline` as it's mis-named for "not high". This will be addressed in a follow-up change.

Canonical link: https://commits.webkit.org/297866@main