[visionOS] Converted APMP content fails to enter immersive mode rdar://155465450 https://bugs.webkit.org/show_bug.cgi?id=296484
Reviewed by Andy Estes.

Certain non-APMP videos, when converted to APMP formats, will contain both
immersive/APMP metadata, and also spatial metadata. When both immersive and
spatial metadata are present, we hit a logical inconsistency in our handling
of video entities; we fail to create a spatial entity (because of the presence
of immersive metadata), and we fail to create a default entity (because of
the presence of spatial metadata).

Resolve the inconsistency inside LinearMediaPlayer.makeDefaultEntity()
by checking for the presence of immersive metadata before returning a (nil)
spatial entity, and add more logging to that method as well as
LinearMediaPlayer.maybeCreateSpatialOrImmersiveEntity() so that this
kind of error is more obvious and debuggable in the future.

* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaPlayer.swift:
(WKSLinearMediaPlayer.maybeCreateSpatialOrImmersiveEntity):
(WKSLinearMediaPlayer.maybeClearSpatialOrImmersiveEntity):
(WKSLinearMediaPlayer.makeDefaultEntity):

Canonical link: https://commits.webkit.org/297869@main