[visionOS] Fix race condition between playerIdentifier and videoReceiverTarget with externalPlayback https://bugs.webkit.org/show_bug.cgi?id=296407 rdar://155573225
Reviewed by Andy Estes.

Safari sometimes has missing content in the player in Spatial Browsing. This is
due to a race condition between the playerIdentifier being set and the
videoReceiverTarget being set.

The existing flow is this:
1. Safari calls enterExternalPlayback
2. WebPageProxy kicks off the request to set the playerIdentifier
3. WebPageProxy also asks the controls manager VideoPresentationInterface to
   enter external playback.
4. VPI puts LinearMediaPlayer into `external` state and notifies Safari.
5. Safari adds the LMPlayableViewController to its scene.
6. AVKit gets notified from REKit about this newly created videoReceiverEndpoint.
7. AVKit informs LinearMediaPlayer who informs the PlaybackSessionInterface.
8. The interface attempts to retrieve the playerIdentifier from the model
   and pipe it (together with the endpoint) to the WebGPU process.

The problem with the above flow is that Step 2 needs to be completed before
Step 8, but Step 2 has to roundtrip from the UIP to the WCP and back for Step 8
to work (for the playerIdentifier to be set on the model in the UIP).

And Steps 3-7 do not require sending messages to the WCP and waiting for the
result. There is *some* async aspects (Step 5 -> 6 takes some time), so this has
worked for the most part, but *sometimes* it happens quicker than the player id.

In the logs, this issue looks like:
17:12:13.397 PlaybackSessionManagerProxy::setVideoReceiverEndpoint(CF500F83) no player identifier
17:12:13.475 WebContent[1270] HTMLMediaElement::setPlayerIdentifierForVideoElement(CD087073AA0B5226)

To fix this race condition, we move Step 2 into Step 4, and VPI does not call the
handler for Safari until after the player identifier is set. Because this introduces
a non-synchronous flow from when VPI attempts to enter external playback to when it
completes, we add a new state on LinearMediaPlayer:

`enteringExternal`

This is to allow for a "lock" on the state so that other states (like fullscreen)
do not interfere (and the reverse) during this transitional period while VPI is
waiting for the playerIdenfitifer to be set. It also allows cleanup to work if
the video were to be cleaned up during this brief period of time.

* Source/WebCore/platform/ios/PlaybackSessionInterfaceIOS.mm:
(WebCore::PlaybackSessionInterfaceIOS::setPlayerIdentifier):
* Source/WebCore/platform/ios/VideoPresentationInterfaceIOS.h:
(WebCore::VideoPresentationInterfaceIOS::didSetPlayerIdentifier):
* Source/WebKit/Platform/ios/PlaybackSessionInterfaceLMK.mm:
(WebKit::PlaybackSessionInterfaceLMK::supportsLinearMediaPlayerChanged):
* Source/WebKit/Platform/ios/VideoPresentationInterfaceLMK.h:
* Source/WebKit/Platform/ios/VideoPresentationInterfaceLMK.mm:
(WebKit::VideoPresentationInterfaceLMK::enterExternalPlayback):
(WebKit::VideoPresentationInterfaceLMK::exitExternalPlayback):
(WebKit::VideoPresentationInterfaceLMK::cleanupExternalPlayback):
(WebKit::VideoPresentationInterfaceLMK::didSetPlayerIdentifier):
* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.mm:
(-[WKWebView _enterExternalPlaybackForNowPlayingMediaSessionWithEnterCompletionHandler:exitCompletionHandler:]):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::setPlayerIdentifierForVideoElement): Deleted.
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaPlayer.swift:
(WKSLinearMediaPlayer.enterExternalCompletionHandler):
(WKSLinearMediaPlayer.enterExternalPlayback(_:(any Error)?) -> Void:)):
(WKSLinearMediaPlayer.completeEnterExternalPlayback):
(WKSLinearMediaPlayer.exitExternalPlayback(_:(any Error)?) -> Void:)):
(WKSLinearMediaPlayer.enterFullscreen(_:(any Error)?) -> Void:)):
(WKSLinearMediaPlayer.exitFullscreen(_:(any Error)?) -> Void:)):
(WKSLinearMediaPlayer.presentationStateChanged(_:)):
(WKSLinearMediaPlayer.toggleInlineMode):
(WKSLinearMediaPlayer.willEnterFullscreen):
(WKSLinearMediaPlayer.willExitFullscreen):
(WKSLinearMediaPlayer.enterExternalPresentation(_:(any Error)?) -> Void:)): Deleted.
(WKSLinearMediaPlayer.exitExternalPresentation(_:(any Error)?) -> Void:)): Deleted.
* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaTypes.swift:
(WKSLinearMediaPresentationState.description):
* Source/WebKit/WebKitSwift/LinearMediaKit/WKSLinearMediaPlayer.h:
* Source/WebKit/WebKitSwift/LinearMediaKit/WKSLinearMediaTypes.h:

Canonical link: https://commits.webkit.org/297873@main