[Absolute Positioning][Grid] align-self self-end and flex-end are equivalent to end for simple content https://bugs.webkit.org/show_bug.cgi?id=296439 rdar://156623092
Reviewed by Alan Baradlay.

In 297803@main, we added some support for alignment of boxes within its
static position rectangle when the parent is a grid, it is not the
containing block, and it has align-self: end. However, both flex-end and
self-end are equivalent to align-self: end in these contexts. For
flex-end, this value simply behaves as end. For self-end, since the
writing modes of the box and alignment container should be the same, the
box should get aligned in the same way.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-flex-end-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-flex-end.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-self-end-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-self-end.html: Added.
These tests are equivalent to the tests in the patch mentioned above but
are adjusted to use these new values to make sure everything works.

* Source/WebCore/rendering/PositionedLayoutConstraints.cpp:
(WebCore::PositionedLayoutConstraints::isEligibleForStaticRangeAlignment const):
Expands eligibility to allow ItemPosition values of FlexEnd and SelfEnd
but also a small refactoring so that we switch on the item position and
then add additional constraints instead of looking at item position as
one of the last pieces of criteria. As eligibility expands in future
patches, the current values may diverge to have different constraints
(e.g. when we incorporate writing mode changes, end and self-end may
require different constraints).

Canonical link: https://commits.webkit.org/297875@main